Move bigBrotherParsing tests
